// de-care/libs/data-services/src/lib/enums/package.enum.ts
/**
 * @deprecated Use PackagePlatformEnum from @de-care/domains/offers/state-offers instead
 */
export enum PackagePlatformEnum {
    Xm = 'XM',
    Siriusxm = 'SiriusXM',
    Sirius = 'SIRIUS'
}

export enum PackagePlanTypeEnum {
    Select = 'Select',
    Allaccess = 'All Access',
    MostlyMusic = 'Mostly Music',
    AllAccessFamilyFriendly = 'All Access Family Friendly',
    SelectFamilyFriendly = 'Select Family Friendly',
    NewsSportsAndTalk = 'News, Sports and Talk',
    None = ''
}